Add files via upload
In this updated version I have used Lasso regression to identify the best feature. And the best feature was the origin of the cars. It was fairly evident from all the scatterplots based on the origin, that cars from the USA stood out from the rest in almost all the available parameters.